backend/pyproject.toml
[tool.poetry]
name = "patent-search-backend"
version = "0.1.0"
description = "FastAPI backend for patent search tool"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
requests = "^2.26.0"
sqlalchemy = "^1.4.23"
psycopg2-binary = "^2.9.1"
pypdf2 = "^3.0.1"
python-multipart = "^0.0.9"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
